moved code to helpers
